add open disk conversation history button
