+++ webview-ui/src/components/chat/task-header/TaskHeader.tsx
@@ -12,6 +12,10 @@ import { VSCodeButton } from "@vscode/webview-ui-toolkit/react"
 import React, { memo, useEffect, useMemo, useRef, useState } from "react"
 import { useWindowSize } from "react-use"
 import TaskTimeline from "./TaskTimeline"
+import DeleteTaskButton from "./buttons/DeleteTaskButton"
+import CopyTaskButton from "./buttons/CopyTaskButton"
+import OpenDiskTaskHistoryButton from "./buttons/OpenDiskTaskHistoryButton"
+
 const { IS_DEV } = process.env
 
 interface TaskHeaderProps {
@@ -412,9 +416,12 @@ const TaskHeader: React.FC<TaskHeaderProps> = ({
 								</div>
 								{!shouldShowPromptCacheInfo() && (
 									<div className="flex items-center flex-wrap">
-										<CopyButton taskText={task.text} />
-										{IS_DEV === '"true"' && <TaskFolderButton taskId={currentTaskItem?.id} />}
-										<DeleteButton taskSize={formatSize(currentTaskItem?.size)} taskId={currentTaskItem?.id} />
+										{IS_DEV === '"true"' && <OpenDiskTaskHistoryButton taskId={currentTaskItem?.id} />}
+										<CopyTaskButton taskText={task.text} />
+										<DeleteTaskButton
+											taskSize={formatSize(currentTaskItem?.size)}
+											taskId={currentTaskItem?.id}
+										/>
 									</div>
 								)}
 							</div>
@@ -468,9 +475,12 @@ const TaskHeader: React.FC<TaskHeaderProps> = ({
 										)}
 									</div>
 									<div className="flex items-center flex-wrap">
-										<CopyButton taskText={task.text} />
-										{IS_DEV === '"true"' && <TaskFolderButton taskId={currentTaskItem?.id} />}
-										<DeleteButton taskSize={formatSize(currentTaskItem?.size)} taskId={currentTaskItem?.id} />
+										{IS_DEV === '"true"' && <OpenDiskTaskHistoryButton taskId={currentTaskItem?.id} />}
+										<CopyTaskButton taskText={task.text} />
+										<DeleteTaskButton
+											taskSize={formatSize(currentTaskItem?.size)}
+											taskId={currentTaskItem?.id}
+										/>
 									</div>
 								</div>
 							)}
@@ -533,34 +543,6 @@ const TaskHeader: React.FC<TaskHeaderProps> = ({
 					</>
 				)}
 			</div>
-			{/* {apiProvider === "" && (
-				<div
-					style={{
-						backgroundColor: "color-mix(in srgb, var(--vscode-badge-background) 50%, transparent)",
-						color: "var(--vscode-badge-foreground)",
-						borderRadius: "0 0 3px 3px",
-						display: "flex",
-						justifyContent: "space-between",
-						alignItems: "center",
-						padding: "4px 12px 6px 12px",
-						fontSize: "0.9em",
-						marginLeft: "10px",
-						marginRight: "10px",
-					}}>
-					<div style={{ fontWeight: "500" }}>Credits Remaining:</div>
-					<div>
-						{formatPrice(Credits || 0)}
-						{(Credits || 0) < 1 && (
-							<>
-								{" "}
-								<VSCodeLink style={{ fontSize: "0.9em" }} href={getAddCreditsUrl(vscodeUriScheme)}>
-									(get more?)
-								</VSCodeLink>
-							</>
-						)}
-					</div>
-				</div>
-			)} */}
 		</div>
 	)
 }
@@ -642,104 +624,5 @@ export const highlightText = (text?: string, withShadow = true) => {
 
 	return [text]
 }
-
-const CopyButton: React.FC<{
-	taskText?: string
-}> = ({ taskText }) => {
-	const [copied, setCopied] = useState(false)
-
-	const handleCopy = () => {
-		if (!taskText) return
-
-		navigator.clipboard.writeText(taskText).then(() => {
-			setCopied(true)
-			setTimeout(() => setCopied(false), 1500)
-		})
-	}
-
-	return (
-		<HeroTooltip content="Copy Task">
-			<VSCodeButton
-				appearance="icon"
-				onClick={handleCopy}
-				style={{ padding: "0px 0px" }}
-				className="p-0"
-				aria-label="Copy Task">
-				<div className="flex items-center gap-[3px] text-[8px] font-bold opacity-60">
-					<i className={`codicon codicon-${copied ? "check" : "copy"}`} />
-				</div>
-			</VSCodeButton>
-		</HeroTooltip>
-	)
-}
-
-const TaskFolderButton: React.FC<{
-	taskId?: string
-}> = ({ taskId }) => {
-	const [copied, setCopied] = useState(false)
-
-	const handleCopy = () => {
-		if (!taskId) return
-
-		navigator.clipboard.writeText(taskId).then(() => {
-			setCopied(true)
-			setTimeout(() => setCopied(false), 1500)
-		})
-	}
-
-	return (
-		<HeroTooltip content="Copy Task ID">
-			<VSCodeButton
-				appearance="icon"
-				onClick={handleCopy}
-				style={{ padding: "0px 0px" }}
-				className="p-0"
-				aria-label="Copy Task ID">
-				<div className="flex items-center gap-[3px] text-[8px] font-bold opacity-60">
-					<i className={`codicon codicon-${copied ? "check" : "folder"}`} />
-				</div>
-			</VSCodeButton>
-		</HeroTooltip>
-	)
-}
-
-const DeleteButton: React.FC<{
-	taskSize: string
-	taskId?: string
-}> = ({ taskSize, taskId }) => (
-	<HeroTooltip content="Delete Task & Checkpoints">
-		<VSCodeButton
-			appearance="icon"
-			onClick={() => taskId && TaskServiceClient.deleteTasksWithIds(StringArrayRequest.create({ value: [taskId] }))}
-			style={{ padding: "0px 0px" }}>
-			<div
-				style={{
-					display: "flex",
-					alignItems: "center",
-					gap: "3px",
-					fontSize: "10px",
-					fontWeight: "bold",
-					opacity: 0.6,
-				}}>
-				<i className={`codicon codicon-trash`} />
-				{taskSize}
-			</div>
-		</VSCodeButton>
-	</HeroTooltip>
-)
-
-// const ExportButton = () => (
-// 	<VSCodeButton
-// 		appearance="icon"
-// 		onClick={() => vscode.postMessage({ type: "exportCurrentTask" })}
-// 		style={
-// 			{
-// 				// marginBottom: "-2px",
-// 				// marginRight: "-2.5px",
-// 			}
-// 		}>
-// 		<div style={{ fontSize: "10.5px", fontWeight: "bold", opacity: 0.6 }}>EXPORT</div>
-// 	</VSCodeButton>
-// )
 
 export default memo(TaskHeader)
