+++ webview-ui/src/components/chat/task-header/TaskTimeline.tsx
@@ -4,7 +4,16 @@ import { ClineMessage } from "@shared/ExtensionMessage"
 import { combineApiRequests } from "@shared/combineApiRequests"
 import { combineCommandSequences } from "@shared/combineCommandSequences"
 import TaskTimelineTooltip from "./TaskTimelineTooltip"
-import { COLOR_WHITE, COLOR_GRAY, COLOR_DARK_GRAY, COLOR_BEIGE, COLOR_BLUE, COLOR_RED, COLOR_PURPLE, COLOR_GREEN } from "./colors"
+import {
+	COLOR_WHITE,
+	COLOR_GRAY,
+	COLOR_DARK_GRAY,
+	COLOR_BEIGE,
+	COLOR_BLUE,
+	COLOR_RED,
+	COLOR_PURPLE,
+	COLOR_GREEN,
+} from "../colors"
 
 // Timeline dimensions and spacing
 const TIMELINE_HEIGHT = "18px"
