+++ webview-ui/src/components/chat/task-header/TaskTimelineTooltip.tsx
@@ -1,6 +1,15 @@
 import React from "react"
 import { ClineMessage } from "@shared/ExtensionMessage"
-import { COLOR_WHITE, COLOR_GRAY, COLOR_DARK_GRAY, COLOR_BEIGE, COLOR_BLUE, COLOR_RED, COLOR_PURPLE, COLOR_GREEN } from "./colors"
+import {
+	COLOR_WHITE,
+	COLOR_GRAY,
+	COLOR_DARK_GRAY,
+	COLOR_BEIGE,
+	COLOR_BLUE,
+	COLOR_RED,
+	COLOR_PURPLE,
+	COLOR_GREEN,
+} from "../colors"
 import { Tooltip } from "@heroui/react"
 
 // Color mapping for different message types
